test: move test from ima core to ima server
